Convert valid and invalid tests over
This adds all the original tests in the valid suite as tests in the new
testfile format.  In each case, we simply pipe the entire test file into
a single file and expect no errors.  Meta data is included for tests
which were previously marked ignored.

For the invalid tests, things were a bit more complicated.  Many of the tests
were actually broken, and needed to be fixed.  Others were missing test methods
meaning they could not be executed, etc.  After a fair amount of manual work,
they are all convered over.

+++ src/test/java/wyc/testing/WhileyCompilerTests.java
@@ -31,10 +31,12 @@ import org.junit.runners.Parameterized.Parameters;
 import wycc.lang.Syntactic;
 import wycc.util.*;
 import wycc.util.testing.TestFile;
+import wyil.interpreter.Interpreter;
 import wyil.lang.WyilFile.Attr.SyntaxError;
 import wyc.lang.WhileyFile;
+import wyc.util.ErrorMessages;
+import wyc.util.ErrorMessages.Message;
 import wyc.util.TestUtils;
-
 
 /**
  * Run through all valid test cases with verification enabled. Since every test
@@ -63,20 +65,27 @@ public class WhileyCompilerTests {
 	 * expectation is that compilation should fail with an error and, hence, the
 	 * test fails if compilation does not.
 	 *
-	 * @param testName
-	 *            Name of the test to run. This must correspond to a whiley
-	 *            source file in the <code>WHILEY_SRC_DIR</code> directory.
+	 * @param testName Name of the test to run. This must correspond to a whiley
+	 *                 source file in the <code>WHILEY_SRC_DIR</code> directory.
 	 */
 	protected void runTest(Trie path) throws IOException {
 		Path srcDir = Path.of(WHILEY_SRC_DIR);
 		TestFile tf = readTestFile(srcDir.toFile(), path);
+		// Check whether this test should be ignored or not.
+		boolean ignored = tf.get(Boolean.class, "build.whiley.ignore").orElse(false);
+		boolean strict = tf.get(Boolean.class, "build.whiley.strict").orElse(false);
+		// Yes, test file indicates it should be ignored (for whatever reason).
+		Assume.assumeTrue("Test " + path + " skipped", !ignored);
+		// NOTE: if we get here, then ignored == false
 		Path testDir = srcDir.resolve(path.toPath());
 		forceDelete(testDir);
 		testDir.toFile().mkdirs();
 		int index = 0;
 		HashMap<Trie, TextFile> state = new HashMap<>();
 		for (TestFile.Frame f : tf) {
-			MailBox.Buffered<SyntaxError> handler = new MailBox.Buffered<SyntaxError>();
+			// Filter available markers (ignoring those requiring verification)
+			TestFile.Error[] markers = filterStaticMarkers(f.markers);
+			MailBox.Buffered<SyntaxError> handler = new MailBox.Buffered<>();
 			// Construct frame directory
 			Path frameDir = testDir.resolve("_" + index);
 			index++;
@@ -85,7 +94,7 @@ public class WhileyCompilerTests {
 			mirror(state, frameDir);
 			// Configure compiler
 			wyc.Compiler wyc = new wyc.Compiler().setWhileyDir(frameDir.toFile()).setWyilDir(frameDir.toFile())
-					.setTarget(path).setErrorHandler(handler);
+					.setTarget(path).setErrorHandler(handler).setStrict(strict);
 			// Add source files
 			for (Trie sf : state.keySet()) {
 				sf = Trie.fromString(sf.toString().replace(".whiley", ""));
@@ -94,21 +103,85 @@ public class WhileyCompilerTests {
 			// Check whether build succeeded
 			boolean compiled = wyc.run();
 			// Check whether build should have succeeded
-			boolean shouldCompile = (f.markers.length == 0);
+			boolean shouldCompile = (markers.length == 0);
 			// Interpreter what happened
 			if (compiled && shouldCompile) {
+				TestFile.Error[] runtime_markers = filterRuntimeMarkers(f.markers);
+				TestFile.Error[] actual = new TestFile.Error[0];
 				// Test was expected to compile, so attempt to run the code.
 				String unit = tf.get(String.class, "main.file").orElse("main");
-				TestUtils.execWyil(frameDir.toFile(), path, Trie.fromString(unit));
+				try {
+					TestUtils.execWyil(frameDir.toFile(), path, Trie.fromString(unit));
+				} catch (Interpreter.RuntimeError e) {
+					actual = new TestFile.Error[1];
+					actual[0] = toError(state,e);
+				}
+				compareReportedErrors(actual, runtime_markers);
 			} else if (compiled) {
 				fail("Test should not have compiled!");
 			} else if (!compiled && !shouldCompile) {
 				TestFile.Error[] actual = handler.stream().map(se -> toError(state, se)).toArray(TestFile.Error[]::new);
-				compareReportedErrors(actual, f.markers);
+				compareReportedErrors(actual, markers);
 			} else {
+				// Report errors
+				for (SyntaxError syserr : handler) {
+					TestFile.Error err = toError(state, syserr);
+					TextFile sf = state.get(err.getFilename());
+					System.out.println("error: " + syserr.getMessage());
+					printLineHighlight(System.out, err.getLocation(), sf);
+				}
 				fail("Test should have compiled!");
 			}
 		}
+		// Finally clean up if we get here. Otherwise, leave files in place for
+		// debugging.
+		forceDelete(testDir);
+	}
+
+	private static void printLineHighlight(PrintStream output, TestFile.Coordinate loc, TextFile sourceFile) {
+		// Extract line text
+		String text = sourceFile.getLine(loc.line);
+		// Determine start and end of span
+		int start = loc.range.start;
+		int end = Math.min(text.length() - 1, loc.range.end);
+		// Print out line text
+		output.println(text);
+		// First, mirror indentation
+		String str = "";
+		for (int i = 0; i < start; ++i) {
+			if (text.charAt(i) == '\t') {
+				str += "\t";
+			} else {
+				str += " ";
+			}
+		}
+		// Second, place highlights
+		for (int i = start; i <= end; ++i) {
+			str += "^";
+		}
+		output.println(str);
+	}
+
+	/**
+	 * Filter out markers which cannot be discovered using static checks in WyC. For
+	 * example, markers which can only be detected with more sophisticated
+	 * verification techniques.
+	 *
+	 * @param markers
+	 * @return
+	 */
+	private static TestFile.Error[] filterStaticMarkers(TestFile.Error[] markers) {
+		return Arrays.asList(markers).stream().filter(m -> m.getErrorNumber() < 700).toArray(TestFile.Error[]::new);
+	}
+
+	/**
+	 * Filter out markers which do not arise at runtime.
+	 *
+	 * @param markers
+	 * @return
+	 */
+	private static TestFile.Error[] filterRuntimeMarkers(TestFile.Error[] markers) {
+		return Arrays.asList(markers).stream().filter(m -> m.getErrorNumber() < 716).toArray(TestFile.Error[]::new);
 	}
 
 	/**
@@ -135,19 +208,39 @@ public class WhileyCompilerTests {
 		int start = span.getStart() - l.getOffset();
 		int end = span.getEnd() - l.getOffset();
 		//
-		TestFile.Range range = new TestFile.Range(start,end);
+		TestFile.Range range = new TestFile.Range(start, end);
 		TestFile.Coordinate location = new TestFile.Coordinate(l.getNumber(), range);
 		// Done
-		return new TestFile.Error(errno,filename,location);
+		return new TestFile.Error(errno, filename, location);
 	}
 
-	public static void mirror(Map<Trie,TextFile> state, Path dir) throws IOException {
+	public static TestFile.Error toError(Map<Trie, TextFile> files, Interpreter.RuntimeError err) {
+		int errno = err.getErrorCode();
+		// Identify enclosing source file
+		Trie filename = Trie.fromString(err.getSource().toString() + ".whiley");
+		// Determine the source-file span for the given syntactic marker.
+		Syntactic.Span span = err.getElement().getAncestor(AbstractCompilationUnit.Attribute.Span.class);
+		// Extract source file
+		TextFile sf = files.get(filename);
+		// Extract enclosing line
+		TextFile.Line l = sf.getEnclosingLine(span.getStart());
+		// Convert space into coordinate
+		int start = span.getStart() - l.getOffset();
+		int end = span.getEnd() - l.getOffset();
+		//
+		TestFile.Range range = new TestFile.Range(start, end);
+		TestFile.Coordinate location = new TestFile.Coordinate(l.getNumber(), range);
+		// Done
+		return new TestFile.Error(errno, filename, location);
+	}
+
+	public static void mirror(Map<Trie, TextFile> state, Path dir) throws IOException {
 		for (Map.Entry<Trie, TextFile> e : state.entrySet()) {
 			Path p = dir.resolve(e.getKey().toPath());
 			TextFile f = e.getValue();
 			//
 			p.getParent().toFile().mkdirs();
-			try(FileOutputStream fout = new FileOutputStream(p.toFile())) {
+			try (FileOutputStream fout = new FileOutputStream(p.toFile())) {
 				fout.write(f.getBytes(StandardCharsets.US_ASCII));
 			}
 		}
@@ -161,7 +254,9 @@ public class WhileyCompilerTests {
 				}
 			}
 		}
-		Files.delete(path);
+		if (path.toFile().exists()) {
+			Files.delete(path);
+		}
 	}
 
 	public static TestFile readTestFile(File dir, Trie path) throws IOException {
@@ -181,14 +276,14 @@ public class WhileyCompilerTests {
 		// Implement a greedy diff
 		int ei = 0;
 		int ai = 0;
-		while(ei < expected.length && ai < actual.length) {
+		while (ei < expected.length && ai < actual.length) {
 			TestFile.Error eith = expected[ei];
 			TestFile.Error aith = actual[ai];
 			int c = eith.compareTo(aith);
-			if(c == 0) {
+			if (c == 0) {
 				ei = ei + 1;
 				ai = ai + 1;
-			} else if(c < 0) {
+			} else if (c < 0) {
 				ei = ei + 1;
 				failed = true;
 				System.out.println(">>> " + eith);
@@ -202,17 +297,17 @@ public class WhileyCompilerTests {
 		failed |= (ei < expected.length);
 		failed |= (ai < actual.length);
 		// Print expected
-		for(;ei < expected.length;ei=ei+1) {
+		for (; ei < expected.length; ei = ei + 1) {
 			TestFile.Error eith = expected[ei];
 			System.out.println(">>> " + eith);
 		}
 		// Print missing
-		for(;ai < actual.length;ai=ai+1) {
-			TestFile.Error aith = expected[ai];
+		for (; ai < actual.length; ai = ai + 1) {
+			TestFile.Error aith = actual[ai];
 			System.out.println("<<< " + aith);
 		}
 		//
-		if(failed) {
+		if (failed) {
 			fail("incorrect errors reported");
 		}
 	}
@@ -224,6 +319,7 @@ public class WhileyCompilerTests {
 	// Parameter to test case is the name of the current test.
 	// It will be passed to the constructor by JUnit.
 	private final Trie testName;
+
 	public WhileyCompilerTests(String testName) {
 		this.testName = Trie.fromString(testName);
 	}
@@ -237,5 +333,92 @@ public class WhileyCompilerTests {
 	@Test
 	public void valid() throws IOException {
 		runTest(this.testName);
+	}
+
+	public static void main(String[] args) throws IOException {
+		String srcdir = "tests/invalid".replace('/', File.separatorChar);
+		String targetdir = "tests";
+		Collection<Object[]> results = TestUtils.findTestNames(srcdir);
+//		Collection<Object[]> results = new ArrayList<>();
+//		results.add(new Object[]{"DoWhile_Invalid_6"});
+		System.out.println("Found " + results.size() + ".");
+		int count = 827;
+		for (Object[] r : results) {
+			String name = (String) r[0];
+			String filename = name + ".whiley";
+			try(FileInputStream fin = new FileInputStream(srcdir + "/" + filename)) {
+				byte[] bytes = fin.readAllBytes();
+				List<TestFile.Error> expected = determineExpectedErrors(srcdir + "/" + name + ".sysout",count);
+				String contents = new String(bytes);
+				// Construct test name with padding.
+				String testname = String.format("%1$6s", Integer.toString(count)).replace(" ", "0");
+				FileOutputStream fout = new FileOutputStream(targetdir + "/" + testname + ".test");
+				PrintStream pout = new PrintStream(fout);
+				pout.println("original.name=\"" + name + "\"");
+				pout.println("======");
+				pout.println(">>> main.whiley");
+				pout.println(contents);
+				pout.println("---");
+				// Print out expected errors
+				for (TestFile.Error e : expected) {
+					pout.println("E" + e.getErrorNumber() + " " + e.getFilename() + " " + e.getLocation());
+				}
+				pout.flush();
+				fout.flush();
+				fout.close();
+				count = count + 1;
+			}
+		}
+	}
+
+	public static List<TestFile.Error> determineExpectedErrors(String filename, int count) throws IOException {
+		ArrayList<TestFile.Error> errs = new ArrayList<>();
+		try {
+			try (FileInputStream fin = new FileInputStream(filename)) {
+				String[] lines = new String(fin.readAllBytes()).split("\n");
+				int i = 0;
+				while (i < lines.length) {
+					i = parseError(i, lines, errs);
+				}
+			}
+		} catch (Exception e) {
+			System.out.println("*** ERROR PARSING " + filename + "(=> test " + count + ")");
+		}
+		return errs;
+	}
+
+	public static int parseError(int index, String[] lines, List<TestFile.Error> errs) {
+		String[] split = lines[index].split(":");
+		int line = Integer.parseInt(split[1]);
+		int errno = guessErrorNumber(split[2].substring(1));
+		index = index + 1;
+		while (lines[index].startsWith("\tfound")) {
+			index = index + 1;
+		}
+		index = index + 1; // skip line highlighted
+		String highlight = lines[index];
+		index = index + 1; // skip highlight
+		int start = highlight.indexOf('^');
+		int end = highlight.lastIndexOf('^');
+		TestFile.Range range = new TestFile.Range(start, end);
+		TestFile.Coordinate loc = new TestFile.Coordinate(line,range);
+		errs.add(new TestFile.Error(errno, Trie.fromString("main.whiley"), loc));
+		return index;
+	}
+
+	public static int guessErrorNumber(String msg) {
+		int base = 0;
+		for(Message[] group : ErrorMessages.ERROR_MESSAGES) {
+			if(group != null) {
+				for(int i=0;i!=group.length;++i) {
+					if(group[i] != null && group[i].match(msg)) {
+						return base + i;
+					}
+				}
+			}
+			base += 100;
+		}
+		//
+		throw new IllegalArgumentException("Did not match " + msg);
 	}
 }
